requirements.txt
aiobotocore==2.5.1
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
botocore==1.29.161
charset-normalizer==3.1.0
configparser==5.3.0
frozenlist==1.3.3
fsspec==2023.6.0
idna==3.4
jmespath==1.0.1
kafka-python==2.0.2
multidict==6.0.4
numpy==1.25.0
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
s3fs==2023.6.0
six==1.16.0
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
